requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
streamlit>=1.38.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
faker>=28.0.0
fpdf2>=2.7.0
mcp[cli]>=1.0.0
pypdf>=4.0.0
